+++ CHANGES.md
@@ -2,6 +2,8 @@
 
 ## master branch
 
+## 0.3.0
+
   - Inspects class / module attributes and inserts a stub for each
   - Discovers what exceptions are raised in a function and inserts a stub for each
   - add settings for suppressing autodetection